# docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.LoginPayload:
    properties:
      Email:
        type: string
      Password:
        type: string
    type: object
  model.Auth:
    properties:
      authID:
        type: integer
      createdAt:
        type: string
      deletedAt:
        type: string
      email:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      user_name:
        type: string
    type: object
host: localhost:8083
info:
  contact:
    email: hzy01hzy@gmail.com
    name: HZhenyong
    url: http://www.swagger.io/support
  description: 接口信息
  termsOfService: http://swagger.io/terms/
  title: tool工具接口
  version: "1.0"
paths:
  /docker/diskUsage:
    get:
      description: 检测磁盘
      responses:
        "200":
          description: 磁盘检测通过
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /docker/dockerInfo:
    get:
      description: 获取docker信息
      responses:
        "200":
          description: 获取docker信息成功
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /docker/getVersion:
    get:
      description: 获取docker信息
      responses:
        "200":
          description: 获取版本成功
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /docker/ping:
    get:
      description: 检测心跳
      responses:
        "200":
          description: 心跳检测通过
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /images/getImageInfo:
    get:
      description: 根据镜像id删除镜像
      responses:
        "200":
          description: 删除镜像成功
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /images/getImageList:
    get:
      description: 获取镜像列表
      responses:
        "200":
          description: 获取版本成功
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /images/reTagImage:
    post:
      description: 重命名tag
      responses:
        "200":
          description: 重命名成功
      security:
      - ApiKeyAuth: []
      tags:
      - docker
  /login:
    post:
      description: 用户登录
      parameters:
      - description: Email&Password
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/controller.LoginPayload'
      responses:
        "200":
          description: ""
      tags:
      - auth
  /register:
    post:
      description: 用户注册
      parameters:
      - description: 用户信息
        in: body
        name: auth
        required: true
        schema:
          $ref: '#/definitions/model.Auth'
      responses:
        "201":
          description: 注册成功
      security:
      - ApiKeyAuth: []
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"